+++ server/.prettierrc.json
@@ -1,0 +1,7 @@
+{
+    "trailingComma": "es5",
+    "semi": true,
+    "tabWidth": 2,
+    "singleQuote": true
+  }
+
