+++ server/db.ts
@@ -1,0 +1,22 @@
+import { Pool } from 'pg';
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+export const pool = new Pool({
+  connectionString: process.env.DATABASE_URL,
+  ssl: {
+    rejectUnauthorized: false,
+  },
+  max: 10,
+  idleTimeoutMillis: 30000,
+  connectionTimeoutMillis: 2000,
+});
+
+pool.connect((err: any) => {
+  if (err) {
+    console.error('Error connecting to PostgreSQL database:', err);
+    return;
+  }
+  console.log('Connected to PostgreSQL database!');
+});
